refactoring to make SFTPEngine.init(int requestedVersion) protected
